[ENH]: Lint Rule for "Old-Style" DB Initialization
It was a standard idiom for subclasses to restore the last backup and enable persistence during class initialization. This works fine, unless the project is being loaded as a library to support another project. It that case, the "main" project might want to set up persistence for this depedent project in a different way (e.g. save the data in a different location, or include the data as part of a larger object graph).

To handle the library situation:
- Specify a "main project" via `SpFileDatabase mainProjectName: aString`.
- In DB subclass #initialize, send #setUp instead of explicitly setting up persistence, which will detect if it is the "main project" or in "library mode"