[Enh]: Setup with or without Library-Mode Consideration
I have come to finally think that it's just a bad practice to try to automatically turn on SimplePersistence for projects which could either be libraries or end-user apps. If really needed, a Deployment or similar group could be added with a class-side initialize for the DB, but for personal use, I'll just handle it via the SP Browser. This adds `SpFileDatabase class>>#basicSetup` for that purpose, because it will just set up without regard for "library-mode", which I expect to be removed in the near future after discussing with Dale on the MetaC ML.